connector-service/requirements.txt
google-api-python-client
google-auth
pika
